ARM: dts: exynos4412-odroid: adjust regulator voltages
Some voltages / voltage ranges differ from the ones that are
used in the Hardkernel vendor kernel (HVK). Since the HVK
is heavily tested with respect to system stability, use these
values instead.

While the HVK uses a voltage range of 9~15V for BUCK2 (vdd_arm),
we currently keep the range of 8~14V, especially since
exynos-cpufreq only uses operating parameters in this range.

TODO: I'm not sure how well these values work for the
ODROID-X, where the SoC is not an Exynos4412 Prime.